+++ src/if.test.js
@@ -1,20 +1,19 @@
-import { expect } from 'chai'
-import u from '../lib'
+import u from '.'
 
 describe('u.if', () => {
   it('does not update if the predicate is false', () => {
     const object = { a: 0 }
     let result = u.if(false, { b: 1 }, object)
-    expect(result).to.eql(object)
+    expect(result).toEqual(object)
 
     result = u({ x: u.if(false, 1) }, {})
-    expect(result).to.eql({})
+    expect(result).toEqual({})
   })
 
   it('does update if the predicate is true', () => {
     const object = { a: 0 }
     const result = u.if(true, { b: 1 }, object)
-    expect(result).to.eql({ a: 0, b: 1 })
+    expect(result).toEqual({ a: 0, b: 1 })
   })
 
   it('will use the result of a function passed as a predicate', () => {
@@ -22,7 +21,7 @@ describe('u.if', () => {
     const aIsThree = (x) => x.a === 3
     const result = u.if(aIsThree, { b: 1 }, object)
 
-    expect(result).to.eql({ a: 0 })
+    expect(result).toEqual({ a: 0 })
   })
 
   it('can be partially applied', () => {
@@ -37,11 +36,11 @@ describe('u.if', () => {
       object
     )
 
-    expect(result).to.eql({ a: 3 })
+    expect(result).toEqual({ a: 3 })
   })
 
   it('freezes the result', () => {
-    expect(Object.isFrozen(u.if(true, {}, {}))).to.be.true
-    expect(Object.isFrozen(u.if(false, {}, {}))).to.be.true
+    expect(Object.isFrozen(u.if(true, {}, {}))).toBeTruthy()
+    expect(Object.isFrozen(u.if(false, {}, {}))).toBeTruthy()
   })
 })