+++ src/util/splitPath.test.js
@@ -1,22 +1,21 @@
-import { expect } from 'chai'
-import splitPath from '../../lib/util/splitPath'
+import splitPath from './splitPath'
 
 describe('splitPath', () => {
   it('treats a number as a single step path', () => {
     const path = 1
     const result = splitPath(path)
-    expect(result).to.deep.equal(['1'])
+    expect(result).toEqual(['1'])
   })
 
   it('handles arrays', () => {
     const path = ['foo', 'bar', 'x']
     const result = splitPath(path)
-    expect(result).to.equal(path)
+    expect(result).toEqual(path)
   })
 
   it('handles strings separated by dots', () => {
     const path = 'bar.0.y'
     const result = splitPath(path)
-    expect(result).to.deep.equal(['bar', '0', 'y'])
+    expect(result).toEqual(['bar', '0', 'y'])
   })
 })