+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   [#467]https://github.com/substantial/updeep/pull/467)
 - Update readme: Lodash-fp is discontinued, link to lodash/fp instead #470
 - Add check for process for browser context #424 
+- Make array manipulation helpers explicit in ReadMe #125 
 
 ## [1.2.0]