+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Update readme: Lodash-fp is discontinued, link to lodash/fp instead #470
 - Add check for process for browser context #424 
 - Make array manipulation helpers explicit in ReadMe #125 
+- add the two argument curried version of updateIn #287
 
 ## [1.2.0]