+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add check for process for browser context #424 
 - Make array manipulation helpers explicit in ReadMe #125 
 - add the two argument curried version of updateIn #287
+- Move from Chai to Jest.
 
 ## [1.2.0]